+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Refactored internal dns-resolver
+* Added option `config.WithInternalDNSResolver` for use internal dns-resolver and use resolved IP-address for dialing instead FQDN-address
+
 ## v3.15.1
 * Removed all conditions for trace retry errors
 * Fixed background color of warn messages